Get rid of HttpClientModule from components
