+++ src/core/integration.ts
@@ -1,17 +1,16 @@
 
-import { NgModule, Inject, NgZone } from '@angular/core';
+import { NgModule, Inject, NgZone, Optional } from '@angular/core';
+import { XhrFactory } from '@angular/common/http';
+import * as httpRequest from 'devextreme/core/http_request';
 import { DOCUMENT } from '@angular/common';
-import { NgHttp } from './http-request';
 
 import * as domAdapter from 'devextreme/core/dom_adapter';
 import * as readyCallbacks from 'devextreme/core/utils/ready_callbacks';
 
 const events = ['mousemove', 'mouseover', 'mouseout', 'scroll'];
-@NgModule({
-    providers: [ NgHttp ]
-})
+@NgModule({})
 export class DxIntegrationModule {
-    constructor(@Inject(DOCUMENT) document: any, ngZone: NgZone, _ngHttp: NgHttp) {
+    constructor(@Inject(DOCUMENT) document: any, ngZone: NgZone, @Optional() xhrFactory: XhrFactory) {
         domAdapter.inject({
             _document: document,
 
@@ -37,6 +36,20 @@ export class DxIntegrationModule {
 
             isDocument: function(element) {
                 return element && element.nodeType === 9;
+            }
+        });
+        
+        httpRequest.inject({
+            getXhr: function() {
+                if(!xhrFactory) {
+                    return this.callBase.apply(this);
+                }
+                let _xhr = xhrFactory.build();
+                if (!('withCredentials' in _xhr)) {
+                    _xhr['withCredentials'] = false;
+                }
+
+                return _xhr;
             }
         });
